requirement.txt
python-dotenv
openai
pyupbit